requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.7
prisma==0.12.0

# Authentication and Security
cryptography==41.0.7
PyJWT==2.8.0

# Utilities and Image Processing
python-decouple==3.8
Pillow==10.0.1
python-dotenv==1.0.0

# Development and Debugging
django-debug-toolbar==4.2.0
drf-spectacular==0.26.5

# HTTP and API
requests==2.31.0
urllib3==2.0.7

# Date and Time
python-dateutil==2.8.2

# Environment and Configuration
gunicorn==21.2.0

# Additional Dependencies
asgiref==3.7.2
sqlparse==0.4.4

# Excel Support
openpyxl==3.1.2
 
 